Added basic webscraper call order and moved the example to FredMeyerScraper
